tfm: Fix flash region and layout for TF-M firmware upgrade
Fix flash region and layout for TF-M firmware upgrade using
MCUBoot instead of BL2 as the bootloader.
Remove the BL2 partition from partition manager.
Define BL2 in TF-M in order to compile in support for a bootloader,
needed in order to set proper permissions for the secondary partition.

Add compilation errors for unsupported use-cases:
- Only MCUBOOT_IMAGE_NUMBER 1 is supported.
- Execute from secondary partition is not supported.

NCSDK-11870

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>