drivers: entropy: Remove for TF-M
Now Zephyr supports a PSA entropy driver
which should be used when TF-M is enabled.
So remove the logic for TF-M in the CC3XX
entropy driver.

Ref: NCSDK-14509

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>

drivers: entropy: Remove CC3XX for TF-M/non-SPM

Removes the CC3XX entropy driver for TF-M and secure
only targets. The TF-M target (using the PSA APIs for
entropy) is now supported by Zephyr. The entropy
driver for secure only targets will be moved to
Zephyr for better intergration with their entropy
drivers.

Ref: NCSDK-14509

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>