Kconfig.nrf: Override TEST_RANDOM_GENERATOR
Zephyr provides a TEST_RANDOM_GENERATOR option
which is dependent on the HAS_ENTROPY not
being enabled. This overrides the Kconfig
option to remove this dependency as our
CC3XX entropy driver is not part of the Zephyr
tree and cannot be included there. Before
this the test random generator never worked
anyway, we always used our hardware RNG
whenever we can. This just removes the build
time Kconfig warning when the test random
generator is enabled.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>