nrf_security: Fix PSA configuration include guard collision
Fix an issue where the config_psa.h header from Oberon PSA core used the
same include guard as nrf-config.h
Because of this the contents of config_psa.h was never included, but
the file itself is always included from build_info.h.
We don't want the Oberon PSA core version of this file because it
defines rules for mbedtls builtin PSA driver, which is not supported
in nrf_security.

Fix this by introducing an empty file that removes the rules, which are
already covered in nrf-config.h, and change the include guard of
nrf-config.h

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>